renamed to dropdown-icons
